+++ app/pages/Contacts/ContactCard.view/ContactCard.view.tsx
@@ -4,8 +4,8 @@ import type { FC } from 'react';
 import { Avatar, Card, CardActionArea, CardHeader, Grid, makeStyles } from '@material-ui/core';
 import ChevronRightIcon from '@material-ui/icons/ChevronRight';
 
-import { StarIcon } from '../../components/icons';
-import type { Theme } from '../../theme';
+import { StarIcon } from '../../../components/icons';
+import type { Theme } from '../../../theme';
 import { ContactCardProps } from './ContactCard.d';
 
 const useStyles = makeStyles((theme: Theme) => ({
