+++ app/pages/Contacts/ContactCard.view/index.ts
@@ -1,0 +1,3 @@
+import xx from './ContactCard.view';
+
+export default xx;
