+++ app/pages/Contacts/ContactForm.view/index.ts
@@ -1,0 +1,3 @@
+import xx from './ContactForm.view';
+
+export default xx;
