+++ app/pages/Contacts/ContactsList.view/ContactsList.view.tsx
@@ -14,9 +14,9 @@ import {
 } from '@material-ui/core';
 import { useTranslation } from 'react-i18next';
 
-import { SearchIcon } from '../../components/icons';
-import type { Theme } from '../../theme';
-import { ContactCard } from './ContactCard.view';
+import { SearchIcon } from '../../../components/icons';
+import type { Theme } from '../../../theme';
+import ContactCard from '../ContactCard.view';
 import { ContactsListProps } from './ContactsList';
 
 const useStyles = makeStyles((theme: Theme) => ({
@@ -115,4 +115,3 @@ const ContactsList: FC<ContactsListProps> = ({
 };
 
 export default ContactsList;
-export { ContactsList };
