+++ app/pages/Contacts/ContactsList.view/index.ts
@@ -1,0 +1,3 @@
+import xx from './ContactsList.view';
+
+export default xx;
