+++ package.json
@@ -197,6 +197,7 @@
     "@types/nprogress": "^0.2.0",
     "@types/qrcode.react": "^1.0.1",
     "@types/react": "^17.0.0",
+    "@types/react-color": "^3.0.4",
     "@types/react-dom": "^17.0.0",
     "@types/react-helmet": "^6.1.0",
     "@types/react-redux": "^7.1.9",
